# Coalesce - fixes #15