Still work-in-progress, but done enough to PR.

Writes on the client and server side are batched if possible. The performance improvements are really, really large.

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/tinylib/synapse/17)
<!-- Reviewable:end -->