@ttacon 2015-01-15
Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/tinylib/synapse/17)







---

<sup>**[client.go, line 266 \[r1\]](https://reviewable.io:443/reviews/tinylib/synapse/17#-JfhwPv1vXAuu_dVGn_M)** ([raw file](https://github.com/tinylib/synapse/blob/423f630875a596d501ff7a8740e8cf6e60bf1450/client.go#L266)):</sup>
Any reason for the stylistic difference between this line and the first case in the following select? (i.e. checking ok val vs if it was nil or not?)

---

<sup>**[client.go, line 300 \[r1\]](https://reviewable.io:443/reviews/tinylib/synapse/17#-JfhwyQd4msEaum6_g3C)** ([raw file](https://github.com/tinylib/synapse/blob/423f630875a596d501ff7a8740e8cf6e60bf1450/client.go#L300)):</sup>
I might be missing something, any reason this isn't != 0?

---


<!-- Sent from Reviewable.io -->