@philhofer 2015-01-15
Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/tinylib/synapse/17)







---

<sup>**[client.go, line 266 \[r1\]](https://reviewable.io:443/reviews/tinylib/synapse/17#-JfhwPv1vXAuu_dVGn_M)** ([raw file](https://github.com/tinylib/synapse/blob/423f630875a596d501ff7a8740e8cf6e60bf1450/client.go#L266)):</sup>
No good reason. I'll change that.

---

<sup>**[client.go, line 300 \[r1\]](https://reviewable.io:443/reviews/tinylib/synapse/17#-JfhwyQd4msEaum6_g3C)** ([raw file](https://github.com/tinylib/synapse/blob/423f630875a596d501ff7a8740e8cf6e60bf1450/client.go#L300)):</sup>
I agree; that reads better.

---


<!-- Sent from Reviewable.io -->