@philhofer 2015-01-16
I'm mostly happy with this, but I'm still working on figuring out how to close the output channel without racing.