@philhofer 2015-01-17
For now, I'm just using `sync.WaitGroup` for counting handlers. Fortunately, using it doesn't noticeably affect performance.