@philhofer 2015-01-17
So, a summary of final performance diffs, w/ `GOMAXPROCS=2` and `(*testing.Benchmark).SetParallelism(20)`:

| Benchmark | ns/op Before | ns/op After |
|:---------------:|:-----------------:|:---------------:|
| Client Pool | 9616 | 3678 |
| TCP Echo | 10716 | 2243 |
| Unix no-op | 3225 | 1591 | 
| Pipe no-op | 2745 | 1257 |

Hooray!

Moreover, now things get *marginally* (but consistently) faster in the `GOMAXPROCS=4` case, rather than significantly slower.